requirements.txt
flask
flask-sqlalchemy
flask-login
flask-socketio
flask-wtf
scapy
python-dotenv
werkzeug
email-validator
reportlab
pyotp
eventlet
psutil